+++ packages/vendure-plugin-store-credit/test/helpers.ts
@@ -32,13 +32,16 @@ export const WALLET_FIELDS = gql`
     currencyCode
     balance
     metadata
-    adjustments {
-      id
-      amount
-      description
-      mutatedBy {
+    adjustments(options: $options) {
+      items {
         id
+        amount
+        description
+        mutatedBy {
+          id
+        }
       }
+      totalItems
     }
   }
 `;
@@ -47,19 +50,16 @@ export const GET_WALLET_WITH_ADJUSTMENTS = gql`
   query GetWalletById($id: ID!, $options: WalletAdjustmentListOptions) {
     wallet(id: $id) {
       ...WalletFields
-      adjustmentList(options: $options) {
-        items {
-          id
-        }
-        totalItems
-      }
     }
   }
   ${WALLET_FIELDS}
 `;
 
 export const CREATE_WALLET = gql`
-  mutation CreateWallet($input: CreateWalletInput!) {
+  mutation CreateWallet(
+    $input: CreateWalletInput!
+    $options: WalletAdjustmentListOptions
+  ) {
     createWallet(input: $input) {
       ...WalletFields
     }
@@ -68,7 +68,10 @@ export const CREATE_WALLET = gql`
 `;
 
 export const ADJUST_BALANCE_FOR_WALLET = gql`
-  mutation UpdateWallet($input: AdjustBalanceForWalletInput!) {
+  mutation UpdateWallet(
+    $input: AdjustBalanceForWalletInput!
+    $options: WalletAdjustmentListOptions
+  ) {
     adjustBalanceForWallet(input: $input) {
       ...WalletFields
     }
@@ -98,7 +101,10 @@ export const REFUND_PAYMENT_TO_STORE_CREDIT = gql`
 `;
 
 export const GET_CUSTOMER_WITH_WALLETS = gql`
-  query GetCustomerWithWallets($id: ID!) {
+  query GetCustomerWithWallets(
+    $id: ID!
+    $options: WalletAdjustmentListOptions
+  ) {
     customer(id: $id) {
       id
       wallets {